Refactor address tests in coreTest
