+++ core-test/src/test/scala/org/bitcoins/core/protocol/Bech32Test.scala
@@ -8,10 +8,12 @@ import org.bitcoins.core.protocol.script._
 import org.bitcoins.core.util.{Bech32, Bech32Encoding}
 import org.bitcoins.crypto.ECPublicKey
 import org.bitcoins.testkitcore.gen._
+import org.bitcoins.testkitcore.gen.ln.LnInvoiceGen
 import org.bitcoins.testkitcore.util.BitcoinSUnitTest
 import scodec.bits.ByteVector
 
-import scala.util.Success
+import scala.annotation.tailrec
+import scala.util.{Random, Success}
 
 class Bech32Test extends BitcoinSUnitTest {
 
@@ -267,6 +269,81 @@ class Bech32Test extends BitcoinSUnitTest {
   it must "fail to read a bech32m address" in {
     forAll(AddressGenerator.bech32mAddress) { bech32m =>
       assert(Bech32Address.fromStringT(bech32m.toString).isFailure)
+    }
+  }
+
+  it must "split all LN invoices into HRP and data" in {
+    forAll(LnInvoiceGen.lnInvoice) { invoice =>
+      val splitT =
+        Bech32.splitToHrpAndData(invoice.toString, Bech32Encoding.Bech32)
+      assert(splitT.isSuccess)
+    }
+  }
+
+  it must "split all Bech32 addresses into HRP and data" in {
+    forAll(AddressGenerator.bech32Address) { address =>
+      val splitT =
+        Bech32.splitToHrpAndData(address.value, Bech32Encoding.Bech32)
+      assert(splitT.isSuccess)
+    }
+  }
+
+  it must "serialization symmetry" in {
+    forAll(ScriptGenerators.witnessScriptPubKeyV0,
+           ChainParamsGenerator.networkParams) { case ((witSPK, _), network) =>
+      val addr = Bech32Address(witSPK, network)
+      val spk = Bech32Address.fromStringToWitSPK(addr.value)
+      assert(spk == Success(witSPK))
+    }
+  }
+
+  it must "checksum must not work if we modify a char" in {
+    forAll(AddressGenerator.bech32Address) { addr: Bech32Address =>
+      val old = addr.value
+      val rand = Math.abs(Random.nextInt())
+      val idx = rand % old.length
+      val (f, l) = old.splitAt(idx)
+      val replacementChar = pickReplacementChar(l.head)
+      val replaced = s"$f$replacementChar${l.tail}"
+      //should fail because we replaced a char in the addr, so checksum invalid
+      Bech32Address.fromStringT(replaced).isFailure
+    }
+  }
+
+  it must "fail if we have a mixed case" in {
+    forAll(AddressGenerator.bech32Address) { addr: Bech32Address =>
+      val old = addr.value
+      val replaced = switchCaseRandChar(old)
+      //should fail because we we switched the case of a random char
+      val actual = Bech32Address.fromStringT(replaced)
+      actual.isFailure
+    }
+  }
+
+  @tailrec
+  private def pickReplacementChar(oldChar: Char): Char = {
+    val rand = Math.abs(Random.nextInt())
+    val newChar = Bech32.charset(rand % Bech32.charset.size)
+    //make sure we don't pick the same char we are replacing in the bech32 address
+    if (oldChar == newChar) pickReplacementChar(oldChar)
+    else newChar
+  }
+
+  @tailrec
+  private def switchCaseRandChar(addr: String): String = {
+    val rand = Math.abs(Random.nextInt())
+    val idx = rand % addr.length
+    val (f, l) = addr.splitAt(idx)
+    if (l.head.isDigit) {
+      switchCaseRandChar(addr)
+    } else {
+      val middle =
+        if (l.head.isUpper) {
+          l.head.toLower
+        } else {
+          l.head.toUpper
+        }
+      s"$f$middle${l.tail}"
     }
   }
 }
